+++ src/presenters/interfaces/sample.presenter.interface.ts
