+++ src/presenters/sample.presenter.ts
@@ -1,12 +1,12 @@
 import { Injectable } from "@nestjs/common";
 import SampleEntity from "src/entities/sample.entity";
 import { SampleModel } from "src/models/sample.model";
-import ISamplePresenter from "./interfaces/sample.presenter.usecase";
+import ISamplePresenter from "./interfaces/sample.presenter.interface";
 
 @Injectable()
 export default class SamplePresenter implements ISamplePresenter {
-  convertToEntity (request: SampleModel): SampleEntity {
-    const response : SampleEntity = {
+  convertToEntity(request: SampleModel): SampleEntity {
+    const response: SampleEntity = {
       id: request.id,
       name: request.name,
       age: request.age,
