+++ src/usecases/sample/getAllSample.service.ts
@@ -1,7 +1,7 @@
 import { Inject, Injectable } from "@nestjs/common";
 import SampleEntity from "src/entities/sample.entity";
 import { SampleModel } from "src/models/sample.model";
-import ISamplePresenter from "src/presenters/interfaces/sample.presenter.usecase";
+import ISamplePresenter from "src/presenters/interfaces/sample.presenter.interface";
 import SamplePresenter from "src/presenters/sample.presenter";
 import ISampleRepository from "src/repositories/interfaces/sample.repository.interface";
 import SampleRepository from "src/repositories/sample.repository";
@@ -19,7 +19,6 @@ export default class GetAllSampleService implements GetAllSampleUsecase {
 
   async execute(): Promise<SampleEntity[]> {
     const sampleModel: SampleModel[] = await this.SampleRepository.getAllSample();
-    console.log(sampleModel);
     const response: SampleEntity[] = sampleModel.map(x => this.SamplePresenter.convertToEntity(x));
     return response;
   }
